+++ src/main/java/net/geforcemods/securitycraft/blocks/SecretCeilingHangingSignBlock.java
@@ -20,13 +20,16 @@ import net.minecraft.world.phys.BlockHitResult;
 import net.neoforged.neoforge.common.NeoForge;
 
 public class SecretCeilingHangingSignBlock extends CeilingHangingSignBlock {
+	private final float destroyTimeForOwner;
+
 	public SecretCeilingHangingSignBlock(BlockBehaviour.Properties properties, WoodType woodType) {
-		super(woodType, properties);
+		super(woodType, OwnableBlock.withReinforcedDestroyTime(properties));
+		destroyTimeForOwner = OwnableBlock.getStoredDestroyTime();
 	}
 
 	@Override
 	public float getDestroyProgress(BlockState state, Player player, BlockGetter level, BlockPos pos) {
-		return BlockUtils.getDestroyProgress(super::getDestroyProgress, state, player, level, pos);
+		return BlockUtils.getDestroyProgress(super::getDestroyProgress, destroyTimeForOwner, state, player, level, pos);
 	}
 
 	@Override