+++ src/main/java/net/geforcemods/securitycraft/blocks/reinforced/ReinforcedSnowyDirtBlock.java
@@ -5,6 +5,7 @@ import java.util.function.BiConsumer;
 import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.api.IReinforcedBlock;
 import net.geforcemods.securitycraft.api.OwnableBlockEntity;
+import net.geforcemods.securitycraft.blocks.OwnableBlock;
 import net.geforcemods.securitycraft.misc.OwnershipEvent;
 import net.geforcemods.securitycraft.util.BlockUtils;
 import net.minecraft.core.BlockPos;
@@ -29,15 +30,17 @@ import net.neoforged.neoforge.common.NeoForge;
 
 public class ReinforcedSnowyDirtBlock extends SnowyDirtBlock implements IReinforcedBlock, EntityBlock {
 	private Block vanillaBlock;
+	private final float destroyTimeForOwner;
 
 	public ReinforcedSnowyDirtBlock(BlockBehaviour.Properties properties, Block vB) {
-		super(properties);
+		super(OwnableBlock.withReinforcedDestroyTime(properties));
 		this.vanillaBlock = vB;
+		destroyTimeForOwner = OwnableBlock.getStoredDestroyTime();
 	}
 
 	@Override
 	public float getDestroyProgress(BlockState state, Player player, BlockGetter level, BlockPos pos) {
-		return BlockUtils.getDestroyProgress(super::getDestroyProgress, state, player, level, pos);
+		return BlockUtils.getDestroyProgress(super::getDestroyProgress, destroyTimeForOwner, state, player, level, pos);
 	}
 
 	@Override